requirements.txt
Django==5.2.9

# rest auth
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

django-environ==0.12.0

# cors
django-cors-headers==4.9.0

# ws
channels==4.3.2
channels-redis==4.3.0

# db
psycopg2-binary==2.9.11

# media
pillow==12.0.0

black==25.11.0
